Fix type issue

+++ mtbs_fire_analysis/pipeline/m02_nlcd_mode.py
@@ -1,6 +1,3 @@
-import glob
-from pathlib import Path
-
 import raster_tools as rts
 
 from mtbs_fire_analysis.pipeline.paths import (
@@ -17,7 +14,7 @@ from mtbs_fire_analysis.utils import (
 def main():
     nlcd_paths = [get_nlcd_raster_path(y) for y in range(1984, 2023)]
     stack_rasters_as_vrt(nlcd_paths, NLCD_STACK_VRT_PATH)
-    nlcd_stack = rts.Raster(NLCD_STACK_VRT_PATH)
+    nlcd_stack = rts.Raster(NLCD_STACK_VRT_PATH).astype("int16")
     mode = rts.general.local_stats(nlcd_stack, "mode")
     protected_raster_save_with_cleanup(mode, NLCD_MODE_RASTER_PATH)
